+++ test/unit/sonarqube_code_coverage.sh
@@ -10,6 +10,7 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/call_func.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/create_data_config.py
+coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/create_header.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_mem_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_lag.py
 coverage run -a --source=mongo_rep_admin test/unit/mongo_rep_admin/chk_rep_stat.py